Add no_proxy information to the README
This changeset adds a few extra details and instructions to the egress proxy configuration to address situations where developers have more than one application running within the platform.

Oftentimes, these applications will need to speak with each other (e.g., an API and a client), and there are secure and efficient ways to do that with the `apps.internal` domain routing via network policy configuration.  However, with a proxy in front of the app, that traffic will not get through unless you explicitly configure your apps with the `no_proxy` environment variable and set it to include `apps.internal` as a domain to ignore.

This will allow internal traffic to flow as expected between your applications.